requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
# openai>=0.27.0 # No longer needed directly
python-dotenv>=0.19.0
# aiohttp>=3.8.0 # Can potentially be removed if only httpx is used
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
email-validator>=1.1.3
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
# httpx>=0.23.0 # Duplicate below
python-json-logger>=2.0.0
redis>=4.0.0  # For rate limiting in production
sqlalchemy>=1.4.0  # For database storage
alembic>=1.7.0  # For database migrations
# tenacity>=8.0.0  # Duplicate below
prometheus-client>=0.14.0  # For metrics
sentry-sdk>=1.5.0  # For error tracking

# Added for OpenRouter client
httpx>=0.23.0
tenacity>=8.0.0

# Added for PDF support
pymupdf>=1.23.0

# Added for CLI
typer[all]>=0.9.0

# Added for Rich CLI output
rich>=13.0.0

# --- LangChain Dependencies Removed --- 

# Environment variables should be in .env file
# See .env.example for required configuration
